Added check if zip part exists before deleting comments and vml drawings

+++ src/EPPlus/ExcelWorksheet.cs
@@ -2602,7 +2602,10 @@ namespace OfficeOpenXml
                     if (_comments.Uri != null)
                     {
                         Part.DeleteRelationship(_comments.RelId);
-                        _package.ZipPackage.DeletePart(_comments.Uri);
+                        if (_package.ZipPackage.PartExists(_comments.Uri))
+                        {
+                            _package.ZipPackage.DeletePart(_comments.Uri);
+                        }
                     }
                     if (VmlDrawings.Count == 0)
                     {
@@ -2635,7 +2638,10 @@ namespace OfficeOpenXml
                     if (_vmlDrawings.Part!=null)
                     {
                         Part.DeleteRelationship(_vmlDrawings.RelId);
-                        _package.ZipPackage.DeletePart(_vmlDrawings.Uri);
+                        if (_package.ZipPackage.PartExists(_vmlDrawings.Uri))
+                        {
+                            _package.ZipPackage.DeletePart(_vmlDrawings.Uri);
+                        }
                         DeleteNode($"d:legacyDrawing[@r:id='{_vmlDrawings.RelId}']");
                     }
                 }
